Bug: Class merges to early on missing sub-dependencies.
When a class, which will be loaded in place of a namespace, has not
loaded dependencies, then merged the loader to early and a later
`namespace()` call of the class override the namespace.

To solve it was it necessary, that the execution of the loaded code has
to be done inside of the `namespace` system to be re-execute after the
dependencies where loaded.
To achieve that, was an automatic namespace adding feature necessary.

A cool side effect: The namespace declaraction of the class file are now
optionally and not anymore necessary. That allow easy a reordering by
moving the files in the folder structure.